receiver_app/requirements.txt
azure-servicebus==7.1.1
Flask==1.1.2